move config related stuff into its own module
